requirements.txt
numpy==1.24.3
onnxruntime-gpu==1.15.1
opencv-python==4.8.0.74
Pillow==10.0.0
torch==2.0.1
onnx==1.14.0